Only deploy if we're on master.

+++ buildbot.sh
@@ -8,4 +8,7 @@ bundle install
 ./run_checks.sh
 bundle exec jekyll build
 
-rsync -avc --delete _site/ $SSH_USER@$SSH_HOST:site/
+BRANCH=$(git rev-parse --abbrev-ref HEAD)
+if [[ "$BRANCH" != "master" ]]; then
+  rsync -avc --delete _site/ $SSH_USER@$SSH_HOST:site/
+fi
